Initial Commit
Made legacy TDP application it's own repository.  It is distinctly
different as it is written in Java and uses MDWS to talk to VistA.  It
may be more productive to use it to convert to talk to VIA in lieu of
MDWS or VistA.js with the new Node.js version.